+++ src/api/index.ts
@@ -13,4 +13,5 @@ export * from './block.api';
 export * from './consume-chain.api';
 export * from './returning-flow-rate.api';
 export * from './system.api';
+export * from './verify.api';
 export * from './metadata.api';